+++ NEWS.md
@@ -14,7 +14,7 @@
 
 ## Bugs fixed
 
-* dtime work now on curve dives
+* dtime work now on curve dives ( #1 )
 
 * Correction of additionnal time when using successive dives. Was adding between 4 and 1 minute more than expected when using tables. ( #11 )